feat: Overlap bounded checkpoint offload with independent execution

+++ src/serve/serve_metrics.cpp
@@ -77,7 +77,8 @@ std::string ServeMetrics::render(std::uint32_t max_concurrency,
     append_counter(out, "ninfer:prefix_cache_hit_tokens_total", prefix_cache_hit_tokens_total_);
     append_counter(out, "ninfer:draft_tokens_total", speculative_draft_tokens_total_);
     append_counter(out, "ninfer:draft_accepted_tokens_total", speculative_accepted_tokens_total_);
-    append_counter(out, "ninfer:auto_save_queued_jobs", live.auto_save_queued_jobs);
+    append_counter(out, "ninfer:auto_save_queued_jobs",
+                   static_cast<std::uint64_t>(live.auto_save_queued_jobs));
     append_counter(out, "ninfer:auto_save_queued_bytes", live.auto_save_queued_bytes);
     append_counter(out, "ninfer:auto_save_in_flight_bytes", live.auto_save_in_flight_bytes);
     append_counter(out, "ninfer:auto_save_rejected_jobs_total", live.auto_save_rejected_jobs);
